requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
attrs==23.2.0
beautiful-date==2.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.4.1
gcsa==2.5.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.0
httplib2==0.22.0
idna==3.7
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
py-cord==2.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.1
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.9.4
